GrizzlyBear_RAG/requirements.txt
openai
python-dotenv
langchain
sentence-transformers
numpy
faiss-cpu
transformers
requests
beautifulsoup